+++ starlite/exceptions/__init__.py
@@ -1,0 +1,31 @@
+# flake8: noqa
+from .exceptions import (
+    HTTPException,
+    ImproperlyConfiguredException,
+    InternalServerException,
+    MethodNotAllowedException,
+    MissingDependencyException,
+    NotAuthorizedException,
+    NotFoundException,
+    PermissionDeniedException,
+    ServiceUnavailableException,
+    StarLiteException,
+    TemplateNotFound,
+    ValidationException,
+)
+
+__all__ = [
+    "HTTPException",
+    "ImproperlyConfiguredException",
+    "InternalServerException",
+    "MethodNotAllowedException",
+    "MissingDependencyException",
+    "NotAuthorizedException",
+    "NotFoundException",
+    "PermissionDeniedException",
+    "ServiceUnavailableException",
+    "StarLiteException",
+    "TemplateNotFound",
+    "ValidationException",
+    "utils",
+]