+++ starlite/utils/__init__.py
@@ -9,8 +9,8 @@ __all__ = [
     "create_parsed_model_field",
     "find_index",
     "join_paths",
+    "import_string",
     "is_async_callable",
     "normalize_path",
     "unique",
-    "import_string",
 ]